LayoutTests/fast/forms/get-forms-to-about-blank.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
  <head>
    <script src="../../fast/js/resources/js-test-pre.js"></script>
    <script>
var successfullyParsed = false;

function run() {
    if (window.layoutTestController) {
        window.layoutTestController.dumpAsText();
        window.layoutTestController.waitUntilDone();
    }

    description('Submitting a GET form to &lt;about:blank&gt; should redirect the browser to &lt;about:blank?&gt;.');
    document.getElementById('form').submit();
}

var initialFrameLoaded = false;
var iframe;
function onFrameLoad()
{
    iframe = document.getElementById('iframe');
    if (!initialFrameLoaded) {
        shouldBe('iframe.contentDocument.URL', '"about:blank"');
        initialFrameLoaded = true;
        return;
    }

    shouldBe('iframe.contentDocument.URL', '"about:blank?"');
    isSuccessfullyParsed();
    if (window.layoutTestController)
        window.layoutTestController.notifyDone();
}
    </script>
  </head>
  <body onload="run();">
    <iframe id="iframe" name="iframe" onload="onFrameLoad();" style="display: none;"></iframe>
    <form id="form" action="about:blank" method="GET" target="iframe" style="display: none;"></form>
  </body>
</html>
